+++ src/App.tsx
@@ -6,7 +6,7 @@ import { queryClient } from './network/react-query/queryClient';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import CommunityPage from './features/community/CommunityPage.tsx';
 import WritingPage from './features/writing/WritingPage.tsx';
-import DetailCommunityPage from './features/detailCommunity/DetailCommunityPage.tsx';
+import DetailCommunityPage from './features/community/pages/DetailCommunityPage.tsx';
 
 import ProfileMainPage from './features/profile/pages/ProfileMainPage.tsx';
 import ProfileUserPage from './features/profile/pages/ProfileUserPage.tsx';
