--- src/features/detailCommunity/components/index.ts
@@ -1,5 +1,0 @@
-export { default as WorryContent } from './WorryContent';
-export { default as DetailPostContent } from './DetailPostContent';
-export { default as CommentsContent } from './CommentsContent';
-export { default as Comments } from './Comments';
-export { default as CommentInput } from './CommentInput';
